Fix #60: Fix download notebook to server.
Since 4.3.0, Jupyter requires that a xsrf token is passed with all requests
to the server.

This uses Jupyter's utils.ajax API call to make the request which is a thin
wrapper around jquery.ajax that handles the xsrf authentication.

Downside is that this change requires a minimum version of Jupyter notebook
of 4.3.1